React navigation install - Login
