+++ package.json
@@ -10,8 +10,15 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.2",
     "react": "17.0.2",
     "react-native": "0.68.2",
+    "react-native-gesture-handler": "^2.5.0",
+    "react-native-reanimated": "^2.8.0",
+    "react-native-redash": "^17.0.0",
+    "react-native-safe-area-context": "^4.3.1",
+    "react-native-screens": "^3.13.1",
     "react-native-svg": "^12.3.0"
   },
   "devDependencies": {
